requirements.txt
Django==2.2
psycopg2-binary==2.8.6
djangorestframework==3.11.2
markdown2==2.3.5
django-filter==2.4.0
djangorestframework-simplejwt==5.2.0
